python/requirements.txt
requests
pandas
sqlalchemy
python-dotenv
psycopg2-binary
matplotlib